API cleanup Removing all the unnecessities Smushed and vaporized some redundant recipe stuffs
