+++ src/main/java/wayoftime/bloodmagic/common/recipe/serializer/TartaricForgeRecipeSerializer.java
@@ -15,8 +15,8 @@ import net.minecraft.network.PacketBuffer;
 import net.minecraft.util.JSONUtils;
 import net.minecraft.util.ResourceLocation;
 import net.minecraftforge.registries.ForgeRegistryEntry;
-import wayoftime.bloodmagic.api.SerializerHelper;
-import wayoftime.bloodmagic.api.recipe.RecipeTartaricForge;
+import wayoftime.bloodmagic.recipe.helper.SerializerHelper;
+import wayoftime.bloodmagic.recipe.RecipeTartaricForge;
 import wayoftime.bloodmagic.util.Constants;
 
 public class TartaricForgeRecipeSerializer<RECIPE extends RecipeTartaricForge>
