+++ src/main/java/wayoftime/bloodmagic/common/registration/impl/EntityTypeRegistryObject.java
@@ -5,7 +5,7 @@ import javax.annotation.Nonnull;
 import net.minecraft.entity.Entity;
 import net.minecraft.entity.EntityType;
 import net.minecraftforge.fml.RegistryObject;
-import wayoftime.bloodmagic.api.providers.IEntityTypeProvider;
+import wayoftime.bloodmagic.util.providers.IEntityTypeProvider;
 import wayoftime.bloodmagic.common.registration.WrappedRegistryObject;
 
 public class EntityTypeRegistryObject<ENTITY extends Entity> extends WrappedRegistryObject<EntityType<ENTITY>>
