+++ src/main/java/wayoftime/bloodmagic/core/registry/AlchemyArrayRendererRegistry.java
@@ -11,7 +11,7 @@ import net.minecraft.world.World;
 import net.minecraftforge.api.distmarker.Dist;
 import net.minecraftforge.api.distmarker.OnlyIn;
 import wayoftime.bloodmagic.impl.BloodMagicAPI;
-import wayoftime.bloodmagic.api.recipe.RecipeAlchemyArray;
+import wayoftime.bloodmagic.recipe.RecipeAlchemyArray;
 import wayoftime.bloodmagic.client.render.alchemyarray.AlchemyArrayRenderer;
 
 public class AlchemyArrayRendererRegistry
