+++ src/main/java/wayoftime/bloodmagic/util/providers/IBaseProvider.java
@@ -1,10 +1,10 @@
-package wayoftime.bloodmagic.api.providers;
+package wayoftime.bloodmagic.util.providers;
 
 import net.minecraft.util.ResourceLocation;
 import net.minecraft.util.text.ITextComponent;
 import net.minecraft.util.text.TranslationTextComponent;
-import wayoftime.bloodmagic.api.text.IHasTextComponent;
-import wayoftime.bloodmagic.api.text.IHasTranslationKey;
+import wayoftime.bloodmagic.util.text.IHasTextComponent;
+import wayoftime.bloodmagic.util.text.IHasTranslationKey;
 
 public interface IBaseProvider extends IHasTextComponent, IHasTranslationKey
 {
