+++ src/main/java/wayoftime/bloodmagic/util/text/IHasTranslationKey.java
@@ -1,4 +1,4 @@
-package wayoftime.bloodmagic.api.text;
+package wayoftime.bloodmagic.util.text;
 
 public interface IHasTranslationKey
 {
